misp2elastic/requirements.txt
pymemcache>=2.1.1
requests>=2.9.1
PyYAML>=5.1
